23.LLM/assistant-demo-v2/requirements.txt
python-dotenv~=1.0.1
openai~=1.52.2
requests~=2.31.0
streamlit~=1.39.0
pillow~=10.3.0